requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest-django==4.6.0
coverage==7.3.2 